+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -217,7 +217,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-10-20T12:00:30.324Z",
+    "2025-10-20T13:32:34.206Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -257,7 +257,7 @@ export const CreateDatasetResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-10-20T12:00:30.324Z"))
+  updated: z.date().default(() => new Date("2025-10-20T13:32:34.206Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {