+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -244,7 +244,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K80QWCKN40NQHKM4G6KA0HH1"),
+  _id: z.string().default("01K80X50949Q4YD0Y7B144P26D"),
   display_name: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$inboundSchema,
@@ -288,7 +288,7 @@ export const UpdateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01K80QWCKN40NQHKM4G6KA0HH1"),
+  id: z.string().default("01K80X50949Q4YD0Y7B144P26D"),
   displayName: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$outboundSchema,