+++ packages/orq-rc/RELEASES.md
@@ -4379,3 +4379,13 @@ Based on:
 - [typescript v3.14.0-rc.38] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.38] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.38 - packages/orq-rc
+
+## 2025-10-20 13:34:23
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.637.0 (2.727.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.39] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.39] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.39 - packages/orq-rc